Moved handle post and upload to s3 functions to different files
