+++ src/app/upload/page.jsx
@@ -9,6 +9,7 @@ import Media from "./media";
 import GiphyButton from "./giphybutton";
 import GiphySelector from "./giphyselector";
 import supabaseAnon from "@/lib/supabaseAnonClient";
+import { handlePost } from "@/lib/post";
 import { useRouter } from "next/navigation";
 
 export default function Upload({ open, onClose }) {
@@ -40,49 +41,12 @@ export default function Upload({ open, onClose }) {
   //   console.log("media", media);
   // }, [media]);
 
-
+  
   // Sends a POST request to the server
-  const handlePost = async () => {
+  const handleUploadPost = async (uploadText, mediaList, userId) => {
     const { data: { session } } = await supabaseAnon.auth.getSession();
-    // console.log(session);
-
-    const mediaPaths = await uploadToS3(); // ["media/as234234"]
-    // console.log(JSON.stringify(mediaPaths));
-
-    if (!mediaPaths) {
-      console.error("Error uploading media to S3");
-      return;
-    }
-
-    // Create the form data
-    const formData = new FormData();
-    formData.append("text", text);
-    formData.append("media_path", JSON.stringify(mediaPaths));
-    formData.append("user_id", session.user.id);
-    
-    // Access the url for the GIFS
-    media.forEach((mediaFile) => {
-      if (mediaFile.type === "image/gif") {
-        formData.append("url", mediaFile.url);
-      }
-    })
-
-    try {
-      const response = await fetch("/api/upload", {
-        method: "POST",
-        body: formData,
-      });
-
-      if (!response.ok) {
-        console.error("Error posting");
-      }
-
-      // const result = await response.json();
-      // console.log("Post successful");
-    } catch (error) {
-      console.error("Error posting", error);
-    }
-
+    handlePost(text, media, session.user.id);
+  
     // Redirect to the home page and close the popup
     Router.push('/home');
     onClose();
@@ -90,50 +54,7 @@ export default function Upload({ open, onClose }) {
     // Reset the text and media state
     setText("");
     setMedia([]);
-  }
-
-  // Handles the POST request to the server(S3 Bucket)
-  const uploadToS3 = async () => {
-    const formData = new FormData();
-    const path = "media/";
-    let mediaPaths = [];
-
-    media.map((mediaFile) => {
-      if (mediaFile.file) {
-        // Append file to form data object
-        formData.append("files", mediaFile.file);
-      } 
-    });
-
-    // Append path to form data only if there are files
-    if (formData.has("files")) {
-      // Append path to form data object
-      formData.append("path", path);
-
-      // Send POST request to S3 Bucket
-      try {
-        const response = await fetch("/api/s3", {
-          method: "POST",
-          body: formData,   // if were sending text it would be JSON.stringify()
-        });
-
-        if (response.error) {
-          console.error("Error uploading file to S3");
-        }
     
-        const data = await response.json();
-    
-        // data.data is the array of image paths in s3
-        // console.log(data.data);
-        mediaPaths = [...data.data];
-      } catch (error) {
-        //indicated frontend error
-        console.error(error);
-        return null;
-      }
-    }
-
-    return mediaPaths;
   };
 
   // Create a function to handle the addition of media due to the issue of having multiple alerts when the media limit is reached in the image/video vs gif upload
@@ -220,7 +141,7 @@ export default function Upload({ open, onClose }) {
             <div className="upload-utilities-right-section">
               <button 
                 className="upload-button"
-                onClick={handlePost}
+                onClick={handleUploadPost}
               >
                 Post
               </button>
